[WIP] single request mode

+++ ratemon/runtime/c/ratemon_tc.bpf.c
@@ -256,8 +256,8 @@ int do_rwnd_at_egress(struct __sk_buff *skb) {
               "actual_end=%u",
               flow.local_port, flow.remote_port, actual_grant_end_seq);
     RM_PRINTK("INFO: 'do_rwnd_at_egress' flow %u<->%u grant_done_check: "
-              "end-ack=%u",
-              flow.local_port, flow.remote_port, actual_grant_end_seq-ack_seq);
+              "end-ack=%d",
+              flow.local_port, flow.remote_port, (int)actual_grant_end_seq-(int)ack_seq);
     RM_PRINTK("INFO: 'do_rwnd_at_egress' flow %u<->%u grant_done_check: "
               "should_trigger=%u ",
               flow.local_port, flow.remote_port, should_trigger);
